requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
stripe==7.10.0
httpx==0.26.0
pytest==8.0.0
pytest-asyncio==0.23.5
scikit-learn==1.4.0
joblib==1.3.2
numpy==1.26.3
pandas==2.2.0
